Finished all in one go.
2 tests had to be adjusted because there were problems with the IPs.
One IP was not reachable at all, so you couldnt scan any ports either,
although the test expected an open port.
One other IP was not resolvable. It was not possible at all to get the
hostname. The host did not allow that. But the test expected that to be
allowed. So I changed it.